pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xbmind"
version = "0.1.0"
description = "Turn any Bluetooth speaker into an offline-first AI smart speaker"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "XBMind Contributors"},
]
keywords = ["ai", "smart-speaker", "bluetooth", "voice-assistant", "offline", "tts", "stt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Home Automation",
]
dependencies = [
    "dbus-python>=1.3.2",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "onnxruntime>=1.16.0",
    "faster-whisper>=1.0.0",
    "piper-tts>=1.2.0",
    "ollama>=0.4.0",
    "httpx>=0.27.0",
    "openai>=1.12.0",
    "anthropic>=0.25.0",
    "google-generativeai>=0.5.0",
    "elevenlabs>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "aiosqlite>=0.20.0",
    "aiohttp>=3.9.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
# torch and openwakeword are installed separately by install.sh:
#   - torch: CPU-only build to avoid ~2 GB CUDA download
#   - openwakeword: --no-deps to skip tflite-runtime (no Python 3.13+ wheels)
full = [
    "torch>=2.1.0",
    "openwakeword>=0.6.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
xbmind = "xbmind.main:main"

[project.urls]
Homepage = "https://github.com/xbmind/xbmind"
Documentation = "https://github.com/xbmind/xbmind/tree/main/docs"
Repository = "https://github.com/xbmind/xbmind"
Issues = "https://github.com/xbmind/xbmind/issues"

[tool.setuptools.packages.find]
include = ["xbmind*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["xbmind"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring hardware (deselect with '-m \"not integration\"')",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
